Add preference for py over pyi for LSP gotodef
Summary:
# This diff...

allows for preference of py over pyi in module finding logic when looking for fallbacks. It is useful in the context of gotodef, when we may prefer to send a user to an implementation file instead of the interface.

# This diff stack...

implements preference for `.pyi` files over `.py` over `.pyc` files in import finding, even if the import occurs later in the includes list.

Previous diffs set up for this by:
- creating `.py` and `.pyi` single file variants in `FindResult`
- removing `use_untyped_imports`
- return remaining search roots from `find_one_part`
- create a `FindResult` comparator
- create method to convert `FindResult` to `ModulePath`
- create `find_module_components`
-  use `find_module_components` in `find_module_in_search_path` and `find_module_in_site_package_path`
- combine `find_module_in_search_path` and `find_module_in_site_package_path`

See bottom of diff stack for more details.

Reviewed By: grievejia

Differential Revision: D79593712

fbshipit-source-id: ab06c2f49df4bb5922cabe5a7a19576e3ee3c2b2