Determine if a module is a test file
Summary:
Pysa has a set of heuristics to determine if a module is a test file, because we want to skip the analysis of those for performance reasons.
Given that the heuristic is pretty simple, I think it's worth moving it in Pyrefly, this way we can entirely skip parsing the source file in Pysa, which should speed up performance.

This diff adds the `is_test` flag to modules in the `--report-pysa` option for pyrefly.

Reviewed By: tianhan0

Differential Revision: D79802285

fbshipit-source-id: 4c4c1cd5c51f9f1ee3874b9479c929726b25ee9d