Simplify overload selection
Summary:
This diff reverts the changes in D75272289. An overload that passes with arg errors definitely passes, as any arg errors are unrelated to signature selection.

The change in D75272289 was a fix for #174, however I am not able to reproduce that bug after this change, and indeed all tests continue to pass.

As a further simplification, we can avoid creating a fresh arg errors collector for each overload; we can just use the main error collector. While we might add redundant errors, as arg expressions might be checked multiple times, those errors will be deduplicated.

Reviewed By: yangdanny97

Differential Revision: D79783838

fbshipit-source-id: f0ca8f0185f85cede6f67f36ea2ac8c7491a5710