Pass error collector along with contextual type hint
Summary:
See previous diff for full context. Briefly, when we are contextually typing an expression, if we want to create an error based on mismatch with a hint, we might need to use a different error collector.

This diff introduces Hint and HintRef types. Most places take a HintRef, but we also have Hint to represent "owned" hints, generally arising from `decompose_*` functions defined in unwrap.rs.

This diff just changes all the code from `&Type` to `HintRef` but does not fix the bug documented in the previous diff.

Reviewed By: yangdanny97

Differential Revision: D79746211

fbshipit-source-id: cda4361d10abdff45ac4c88534ea5d4834b62b13