Document missing overload selection failure
Summary:
Overload selection depends on separate error sets, one for inferring argument types, one for matching arguments to parameters.

Contextual typing of arguments based on parameters is part of the latter, so any errors that arise from contextual typing need to be added to the "call" errors in the overloaded call.

One place where this is violated is contextually typing dict literals with a typed dictionary hint. At this point in the expr code, we only have one error collector, and it's the "arg" errors in an onverloaded call.

We could fix this by doing what other contextual typing code paths do -- don't error if we fail to match the hint, but just fall back and infer a `dict[k,v]` type. That would work, but an error specific to typed dictionaries would be better.

Reviewed By: rchen152

Differential Revision: D79746013

fbshipit-source-id: 17d050e18d35a5bb0fbe90e7a80b6f5a7db6ff14