Enforce readonly on pydantic frozen fields
Summary:
Here, we mirror dataclass frozeness check.

- needed to add further basemodel stubs for this

what I should do next is:
model_config is a ClassVariable. I should change my code to specifically match on that.

A more general thought is whether the approach still makes sense/needs more modifications.

One thing that comes to mind is that I need is to know if a field is a ClassVar. For normal fields, we need to add a check to confirm they are annotated, because model fields must all be annotated. That will be the next thing for me to look into.

Reviewed By: rchen152

Differential Revision: D79764759

fbshipit-source-id: 375ffd25added98f3358f022b9dd2ab8652a4029