Parse more pylance options
Summary: In this diff, I start to parse [more settings of pylance](https://code.visualstudio.com/docs/python/settings-reference#_pylance-language-server), so that we can have a more consistent experience wrt pylance out of the box. We don't make use of these settings yet, but in the next few diffs, we will gradually start to make use of them.

Reviewed By: samwgoldman

Differential Revision: D79897288

fbshipit-source-id: 69670eb238ddde290bcc04e70383e84814a1ef0b