Inline the `has_unknown_tparams` calculation into its use site
Summary:
Today there's this `has_unknown_tparams` flag stored in `ClassBases`. Calculation of this flag depends on both solving the full class metadata and solving all the tparams for the given class.

There are many downstream clients for `ClassBases` in general, but there's only 1 downstream client for the `has_unknown_tparams` flag. This means that, instead of storing the flag within `ClassBases` and let all other downstream clients of `ClassBases` unnecessarily depend on the extra work, we could simply calculate the flag at the sole use site of it. This way, we no longer need to solve all tparams when computing `ClassBases`, tidying up the dependency graph and reducing the chance of running into this kind of cycle in prod.

Reviewed By: rchen152

Differential Revision: D79828455

fbshipit-source-id: 63d8f1463973aa78f9d3c5aff01a0c37a67d3fa0