fix narrowing bug with walruses
Summary:
fixes https://github.com/facebook/pyrefly/issues/826

```
def f(x: None | int) -> int:
    if y := x is None:
        x = 1
    return x
```

Heres whats going on:

`y := (x is None)` falls into the "is truthy" path of our narrow generation. This means we only generates a constraint for `y`, since `x is None` isn't a name/attr/subscript. The fix is to also generate narrows for `x is None` as if that was the entire condition.

But now we have two NarrowOps, and we have to merge them. The first one is `IsTruthy(y)` and the second one is `IsNone(x)`

If we use `and_all`, if a name is not present in one NarrowOps it generates a placeholder.

That gives `And(IsTruthy(y), Placeholder(y)); And(IsNone(x), Placeholder(x))`

When negated, this becomes: `Or(IsFalsey(y), Placeholder(y)); Or(IsNotNone(x), Placeholder(x))`, which gives no useful narrowing information because of the placeholders.

Normally this is correct, for example, in a conditional like `if x is None and y` the negation indeed provides no useful information since `x` and `y` are unrelated.

However, in this case the walrus makes it s.t. `y` is only true if `x is None` is also true (and vice versa), so we should merge the entries without any placeholders to be `IsTruthy(y); IsNone(x)`

This makes the negation `IsFalsey(y); IsNotNone(x)`

Reviewed By: samwgoldman

Differential Revision: D79910268

fbshipit-source-id: f9669aa77b86287574d36b4b5a02bd1541fea8d8