fix narrowing bug with or
Summary:
for something like `x or y`, we `or` each entry in the RHS narrow with any matching entry in the LHS narrow (or create an `or` between the RHS entry and a placeholder). we do not add a placeholder for entries in the LHS without a match in the RHS.

So `IsTruthy(x)` OR `IsTruthy(y)` becomes `IsTruthy(x); Or(IsTruthy(y), Placeholder(y))`, making us incorrectly think `x` always truthy.

This diff adds the placeholder for the LHS values.

Reviewed By: stroxler

Differential Revision: D79934516

fbshipit-source-id: 41ada4cb956dcff35e89b5ace3b8840503e53245