Don't require compatible return type for `yield from` (#878)
Summary:
Pull Request resolved: https://github.com/facebook/pyrefly/pull/878

We need to ensure that the `yield` and `send` types are consistent with the outer generator, because send value's are passed in from outer to inner, and yield values from inner to outer. Return values are just returned from the yield from expression itself, so it can have a different type.

Fixes https://github.com/facebook/pyrefly/issues/871

Reviewed By: yangdanny97

Differential Revision: D79909271

fbshipit-source-id: a9aba76ccdcd6e2c3e1f18b58ea4789b7438fa56