Consider types of variables when computing semantic token type for variables
Summary:
Currently, when we see some binding that's syntactically a variable, we will just give it the `variable` kind for semantic highlighting. However, Pylance will actually look at the types of these variables, and provide richer results.

In this diff, we will start doing this as well. We can easily inspect the kind of the type of a given binding that's a variable. For now, I only did a first pass to categorize kinds of types into symbol kinds. Someone who is more familiar with the internal type representation can probably do a bit better than this.

With this and the previous diff, we can close https://github.com/facebook/pyrefly/issues/873

Reviewed By: kinto0

Differential Revision: D79921630

fbshipit-source-id: 72a8cf3675aea6cc8e3c0fe98c23dd75075f298c