Remove short-circuit optimisation
Summary:
Fixes https://github.com/facebook/pyrefly/issues/864

We used to avoid going through answers lookup if we were talking about something in our module, e.g. a class. But you can have both a `.py` and a `.pyi` file in the graph at once, and our rule is that always import from `pyi`. That meant sometimes we reached into the `.py` for things that only existed in the `.pyi` and crashed.

This is a nice optimisation to have, so I've left it commented out. As a warning to those who come later, but also since I imagine there are some conditions that could make it safe.

Reviewed By: yangdanny97

Differential Revision: D80082167

fbshipit-source-id: dc4876a3ade5eaedb6d120371724388758280f3d