Always allow a `...` function parameter default in a stub file
Summary:
Fixes https://github.com/facebook/pyrefly/issues/895.

Technically, type checker behavior is "undefined" for anything except `...` as a function body in a stub file: https://typing.python.org/en/latest/spec/distributing.html#functions-and-methods. But we know that a stub file contains, well, stubs, so it's reasonable to not error on a `...` default regardless.

Reviewed By: ndmitchell

Differential Revision: D80107566

fbshipit-source-id: 35b15ccff2d2f47c8879dbe3b948bd8e8e4f47f8