Traverse the ast with bfs instead of dfs
Summary:
There is currently a bug to find fully qualified names for xrefs when the definition is later in the code. Example:

```
def func():
    def a():
        b()

    def b():
        pass
```

In this case, when calling `b` inside `a`, the fully qualified name is not yet discovered, so the function name defaults to a top-level definition.

To fix this, I am changing the AST traversal from depth-first search (DFS) to breadth-first search (BFS). This approach will also make it easier to find docstrings for variables.

Reviewed By: nhawkes

Differential Revision: D80011346

fbshipit-source-id: cd1ab6ccf81d794e09239553a29a067268a3db75