Use hint for empty list literals
Summary:
This changed, I think unintentionally, in D69789687. It's arguably OK to do this, since the contained var will eventually pin to the elt type when we eventually check against the hint.

However, for soft type hints (added in D80146559), we might not ever check against the hint, so we need to use hint when we push it down.

Reviewed By: ndmitchell

Differential Revision: D79685623

fbshipit-source-id: 96f8f8acf03225facb4d6c2d6cbfe77ad31f97f4