Expose pydantic files to test harness
Summary:
There are various unittests that don't work well since manually writing stubs can be error prone.

We concluded that we need a better approach than manually writing stubs. There are two approaches we considered:

1- expose the entire pydantic library (D80043443)

pros: we don't need to worry about stubs/types being wrong and don't have to modify this frequently

cons: timeout limit needs to increase to 20 seconds so that tests don't timeout when run locally. This seems less of an issue in CI.
Individual tests only take upto 5 seconds, however, when tests are run in parallel, they can take longer, hense the 20 second limit.

2- Only copy the files we need for the unittests and keep adding more, which is the approach in this diff.

pros: This timed out on only one test for me locally, and none for Rebecca, but we increased the timeout limit to 15 seconds to be safe. So it's better than the first approach in terms of speed

cons:  Incomplete pydantic code and we have to go back and keep adding more files, which may introduce a timeout again.

Reviewed By: ndmitchell

Differential Revision: D80121269

fbshipit-source-id: d06e19ea114a238885a72992485d2adb354f0700