Avoid unnecessary clones in is_subset_eq_var
Summary:
The Variable enum is pretty large -- 104 bytes -- mostly because Quantified is pretty large. We were always cloning this value, but never needed to.

We need to clone the types that we use, since those come from the variables lock, which we need to drop before recursing. But we can clone just those types.

Reviewed By: ndmitchell

Differential Revision: D80186509

fbshipit-source-id: 979769a90c4db072271fc0b211db6e5bfadb343b