Simplify is_subset_eq_var
Summary:
In the var-to-var case, if either var is already solved (i.e., Variable::Answer), then we can re-use the existing code for type-to-var and var-to-type.

There is a small difference, although one that doesn't change behavior. We currently allow a var to be "solved" to another var (see unsovlved-var-to-unsolved-var case). Before this change, if we saw a unsolved var1 <: var2 solved to var3, we would solve var1 to var3, even if var3 was solved to a concrete type. After this change, we will end up solving var1 to a concrete type.

This change will also cause us to consume "gas" a bit more quickly than before, although I don't expect this to be a problem in practice, since exhausting gas is a stop-gap to avoid non-termination in extreme scenarios, and we really shouldn't ever get close to it.

Reviewed By: ndmitchell

Differential Revision: D80185289

fbshipit-source-id: 6664efe66f9b26eee89f760bb59dcf4d64bbeaed