Split up determining whether a function decorator is special from the act of applying it to a function
Summary:
This will make it possible for me to determine when we're calling a decorator without inferring argument types for the call. Inferring argument types for abitrary non-decorator functions triggers exponential behavior. This unfortunately requires looping over a function's decorators twice, but this shouldn't be expensive.

Functionally a no-op except for the extra loop.

Reviewed By: samwgoldman

Differential Revision: D80140637

fbshipit-source-id: 1534896af09302dbd2be14bcbff12778ca6879a0