Split up using a decorator to update function flags from checking whether the decorator is valid
Summary:
This will let me update function flags without also doing a validity check. IMO it's also a little more readable to split these up. Now we have to do a third loop over the decorators, but again, this shouldn't be expensive.

Functionally a no-op except for the extra loop.

Reviewed By: samwgoldman

Differential Revision: D80140649

fbshipit-source-id: 14cd2585daf920e71600e3c0e1f89037eed30588