Move pydantic subroutines in their own file
Summary:
We will work on the alias feature next. For this, there will be more parsing of pydantic fields/matching on more keywords.

As the functions get larger, and as Rebecca suggested, we should move them to a different file.

I already have a definitions file, so I thought this is a good place to put them.

Reviewed By: rchen152

Differential Revision: D80201194

fbshipit-source-id: 0f6b7b934e49c12ea232b5bf756f57a4bdbdce01