Move python selector to rightmost side of screen on desktop and bottommost on mobile
Summary:
I finally had some time to take a look at D78350429 after delaying it, and noticed that this was landed (thanks yangdanny97). I was taking a look at the sandbox after and notice a minor issue where the selector was to the left of the buttons on desktop / on top on mobile.

I think it's better visually if the selector is the last button the list (rightmost on desktop / bottommost on mobile) since it's more likely for users to click on the other buttons. The expected workflow is likely to select python version once, but it's possible to need to press run/share/reset multiple times so better to move those closer to the middle of the screen.

Reviewed By: yangdanny97

Differential Revision: D80199442

fbshipit-source-id: 4adc95308c55ced97e8a820115803f2146a844c6