add new argument
Summary:
## Background
Pyrefly currently uses the same definition for go-to definition as typechecking. This is incorrect: for hover, docstrings are generally in the `.py`, not the `.pyi`. For definitions, it's much more useful to see the real implementation. This stack serves to implement "go to definition" correctly.

## Previous approach
previously finder had a naive "preference" that only worked within the same search path. It wouldn’t work when searching subsequent search paths. it wouldn’t work in the following example:
```
typeshed: foo.pyi
site_package: foo.py
```

## This approach
Instead, the approach taken in this stack adds a filter to finder. In order to find the definition of foo.py, we would search for `foo` with a filter on `ModuleStyle::Executable` (represents a `.py`). If that fails, we can fall back to the normal logic.

## This diff
add new argument for file filter. currently unused, will be used in next diff

Reviewed By: connernilsen

Differential Revision: D80145017

fbshipit-source-id: 657887f98976c7598d431eb23edf3b25f8a24dba