Add a regression test for a stub pattern
Summary:
This is a litte bit debateable, because the code is not valid Python
(you can write code that behaves exactly like this in Python, but you
can't do it declaratively - you'd have to insert the subtype classes
into the parent class after the parent class is already defined.

But given that this is how PyO3 represents enums, and we expect use of
PyO3 to increase over time, my opinion is that we should allow it, especially
since it already works today (I'm just adding a regression test so that we know
what breaks, and who would care about it, if we were to change the semantics).

Reviewed By: ndmitchell

Differential Revision: D80093605

fbshipit-source-id: 4b07831db5c55f3b67b9658e5dab1bf80aef6174