fix Self tracking in method bodies
Summary:
Previously we only tracked the type of the `Self` special form if the immediate scope is a class. This meant that usages of `Self` inside the body of a method (which would be in a method scope enclosed by a class scope) did not get tracked.

fixes https://github.com/facebook/pyrefly/issues/815

Reviewed By: rchen152

Differential Revision: D80264585

fbshipit-source-id: a7ae323155a93b744a54b6a5070e6d4b2e24fb62