Don't leak function param var into the static env
Summary:
See D70050192 for context. We create this var because we need to wait until solve to figure out the decorator, to see whether this is a classmethod. To ensure the var is solved before we try to use it, we must solve the FunctionParameter binding to get the type.

Note that this change means we no longer treat un-annotated self/cls bindings as "annotated" targets during assignment. That means the type of self/cls can widen and change. This is a change in behavior, but actually matches Pyright, so I'm just documenting the behavior in a test.

If we want to restore the "unannotated self/cls behaves like it's annotated" behavior, we can do that too. We would need to create a new BindingAnnotation type that gets the type via the FunctionParameter binding, instead of directly using the var.

Reviewed By: ndmitchell

Differential Revision: D80218627

fbshipit-source-id: c86f553063e0290486f216092415b44b51006cc0