Cleanup pydantic frozen implementation
Summary:
I think that the logic inside class_definition for pydantic was too verbose. I wanted to have a single function call that extracts frozen data.

Changes here are:

- replace logic with a single call by passing a mutable reference, since we have context data across fields, and i use a mutable refence to do that
- introduce constants to keep track of keywords

I plan to keep a list of field keywords for aliases, validation_aliases etc. as we need them.

I noticed the dataclass transform mechanism for keeping field keywords was a struct. It's unclear to me yet if I'll need that here.

Right now, I plan on extending the file with more field keyword constants.

Next:
Then the idea will be to make the `self.extract_frozen_pydantic_metadata` function be part of a larger method that also extracts alias data etc.

Reviewed By: ndmitchell

Differential Revision: D80214051

fbshipit-source-id: 39322118b511a7f45f39162de5946310ca8a8bbd