Ensure Binding::FunctionParameter is solved before it is read
Summary:
D77800883 removed the code to force the function binding when solving Binding::FunctionParameter, originally added in D70050192.

It's true that the original code led to cycles, and those cycles caused bad performance and bugs. However, forcing the var like this leads to non-determinism, since the var will get pinned to a type if we solve the function binding first, or any if we try to solve the function param first.

We can eliminate the cycle while solving determinism after D80351861, which introduces the UndecoratedFunction binding. That binding does not depend on the return type, so we can safely force it without introducing a cycle.

Reviewed By: grievejia

Differential Revision: D80224173

fbshipit-source-id: 794b7460d37dd2044848aeefa31d7fd6170f15a8