Remove remaining `mut` variables from `class_metadata_of`
Summary:
Tweaks calculations of bases, is_final, and total_ordering_metadata so that the variables no longer need to be `mut`.

Note that this is a small behavioral change for `total_ordering_metadata` - we previously grabbed the range of the last `total_ordering` decorator; now we grab the range of the first. I don't think this will matter in practice.

Reviewed By: grievejia

Differential Revision: D80372921

fbshipit-source-id: ce25ccf7bd85de763170f7bb64febca736b9d8b6