Adjust display ranges for generator expressions
Summary:
There is a discrepancy between how Python AST and Ruff report locations for generator expressions.

Python AST always wants generator expression to be parenthesized so it always sets col_offset to point to paren, which sometimes does not necessary syntatically belong to generator (i.e when generator is passed as the argument to function and paren is openinng paren in function call)
Ruff distinguished between these two cases and its output has additional attribute `parenthesized` that identifies whether generator expression was originally wrapped in parens. For generators that were not explicitly parenthesized ruff result would have its start column point to the first character of generator which differs from CPython. Given that client will most likely be using Python implementation as a guidance we can conditionally adjust ranges in results of query to be consistent with python ranges.

Reviewed By: brosenfeld

Differential Revision: D80455471

fbshipit-source-id: 3087dd0e24a86a4f505dd3a33f76fdad246e00d4