Simplify some types by exposing trace as owned data, not Arc
Summary:
We have some bugs where we leak too many Vars to the user through traces.

Part of the root cause is that we are overusing `Arc` - to fix the bugs I need
to push down ownership of the type information to a layer where I have enough
information to expland away the Vars.

It turns out the use of `Arc` is pretty silly anyway - the query module (where
we have a bug report coming from) is actually literally the only caller that
*doesn't* immediately clone the arc after getting it anyway, so this is a
zero-behavior-change and near-zero cost simplification.

Reviewed By: rchen152

Differential Revision: D80365201

fbshipit-source-id: cac0c8529ae42960c1fc3fd9f32fa8bb5e91d886