Simplify callee logic just a bit
Summary:
We rely on a lot of nested functions here. Many of them are useful layers, but
one of them is pretty easy to inline, and when we're drowning in layers of indirection
I think it's useful to avoid the layer that isn't just doing much.

This matters especially because the word `target` has no apriori meaning, it's a
made-up term coming from Pyre and Pysa. That means that using lots of abstractions
with "target" in the name doesn't really help much (as opposed to when working with
well-understood abstactions, where a well-named helper function can be quite useful - here,
it just kind of gets in the way).

Reviewed By: rchen152

Differential Revision: D80373707

fbshipit-source-id: 6780a300f65125060874fe05497d79098dc39feb