Rename some mro stuff
Summary:
The name `callee_from_mro` is more consistent with all the other
names.

And what it does (invoke a callback that, given any ancestor, does arbitrary work
to attempt to figure out an applicable method from the ancestor... if that fails
fall back to a default method name that may not actually exist according to Pyrefly,
treating it as defined on the current class) is very poorly expressed unless we
name the arguments well.

Given that the presence of callbacks makes it impossible to even figure out
what this does without cross-referencing the helper against the callback
definitions, that's not so good; renaming a couple of arguments makes it a lot
quicker to see how it's being used (the main use case that requires the full power
of the API is picking a constructor, which is tricky as we know and this code is
trying to do in a roughly-Pyre1-compatible way).

Reviewed By: rchen152

Differential Revision: D80373704

fbshipit-source-id: e14c725ffa672319375b3cf65251988c6855364d