Use a clear name expressing intent
Summary:
This helper is not being used to "get a string from a type". It's
being used to extract a class name from the type that was bound as
the `.obj` field of a `BoundMethod`, which is a very specific operation
that makes perfect sense in context of where it is used.

Let's rename it to make this obvious.

Reviewed By: rchen152

Differential Revision: D80373708

fbshipit-source-id: ce3d19d5460e81a26ff169b05d8caa0d627d4ef4