Push `for_display` calls down into `Answers` API
Summary:
The `for_display` function is really about making a `Type`
meaningful outside of an `Answers` instance:
- expanding the type so that solved vars down't leak out
- simplifying the type if possible
- eliminating Var ids for any vars that remain

The first two operations are unambiguously desirable for absolutely
any downstream use case of data from `Answers`, there's no use case
other than debugging (and we have print debugging + things like
`reveal_type` for that) where you'd want the raw type with possibly
unexpanded Var and unsimplified unions.

The last one is a little less clear, in theory if we wanted to support
downstream users indexing into recursive type cycles we'd need to expose
ids, but I don't envision that happening right now - it would be incredibly
brittle anyway, since our own representation is unstable and will almost
certainly change.

As a result, I believe that this operation should always be done when
exposing a type coming from `Answers` to downstream code, and I'm making
all of the trace-based fetchers use it.

This will fix a bug we've had reported in the query code, where a downstream
tool asking for types of
```
for x in thing:
    mydict[k] = x
```
is seeing `12345` as the type of `mydict` - that's happening because we
put the `for_display` hook in `lsp` thinking it's somehow tied specifically
to hover, but it really belongs in the lower level "get something out of
`Answers`" API that also powers query / pysa / etc, because none of these
tools are going to find an unexpanded var even remotely useful.

(This specific example may change if we do better with loops and speculative
phi, but the point stands that we don't want to leak unexpanded vars or
nondeterministic var ids to downstream tools).

Reviewed By: rchen152

Differential Revision: D80365200

fbshipit-source-id: ba544f62f2cb7714d07d47863949d0b05f7cc2c8