Push chosen overload type construction down
Summary:
I want to fix a var leak bug caused by us eliminating Vars in
lsp, which is too far downstream, whereas we ought to eliminate
them in Answers.

But in order to do that without also introducing inconsistencies in the
API, I also need to expose the chosen overload as a Type instead of a Callable.

The only thing we ever do with it is turn it into a Type anyway, so this is not
a meaningful change, just a refactor to simplify the type signatures at the
API boundaries.

Reviewed By: rchen152

Differential Revision: D80365202

fbshipit-source-id: 8e95f4330a07064702615cca68e8ea2303a4053b