support discriminated unions in narrowing
Summary:
When we have a check like `x.y == 1` or `x["key"] = "value"` or `x[0] is None`, we previously only narrowed the appropriate facet for `x`. This diff makes us also narrow the type of `x`.

This is useful for modeling a form of tagged unions in user code, where we have a union of classes/typeddicts/tuples that's differentiated by a facet that's typed as some literal.

This is a feature that's supported for Pyright, and should be useful for Pydantic users as well.

see: https://github.com/facebook/pyrefly/issues/650

closes https://github.com/facebook/pyrefly/issues/418

Reviewed By: stroxler

Differential Revision: D80464317

fbshipit-source-id: ec583bbb910c53fecff4384b292cf8adc9ea93a9