Push the for_display call down to Answers in get_type
Summary:
This commit is a pure cleanliness tweak; I recently fixed an issue
affecting the query and pysa APIs where - partly as a result of having
pushed too much `for_display` logic up to the lsp layer instead of putting
it in the core Answers api - we were inappropriately exposing raw,
un-expanded types that were useless to clients.

The LSP `get_type` function isn't currently used by any of the query APIs,
so it isn't subject to this problem right now, but it has the same layering
issue where we really shouldn't be exposing raw types that contain
unexpanded variables (which are effectively private data to Answers) in the
API.

This diff adds a helper to layer things better.

Reviewed By: yangdanny97

Differential Revision: D80464998

fbshipit-source-id: abcec486bdca2ac83decc52d8174b1f30565ab45