Do not perform typevar bound checking when calculating class base types
Summary:
When specializing a generic class `Foo[T: A]` with a targ of `B`, we currently does 2 things:
- The "check" part: Attempt to verify that `B` must be a subtype of `A`.
- The "instantiation" part: zip together the targ `B` with the tparam `T`, and arrive at a `ClassType` of `Foo[B]`.

This is all good in normal type checking flow. But when the instantiation happens on the base list, we'll run into the problem of cyclic computation if we do the "check" part (instantiation part is not affected). Consider this example:
```
class A: pass

class Foo[T: A]: pass
class Bar(Foo['Baz']): pass
class Baz(Bar, A): pass
```
This was simplified from the full example in D80222861, but the basic structure of the computation is this:
- We want to calculate the MRO of `Baz`, which depends on the MRO of `Bar`.
- The MRO of `Bar` depends on the MRO, and in turn the base type, of `Foo[Baz]`.
- The base type of `Foo[Baz]` is a generic instantiation, where the targ `Baz` must be validated against the bound of `A`.
- In order to resolve the subtyping request `Baz <: A` we need the MRO of `Baz`. But we are still in the process of computing the MRO of `Baz`! Our current logic will lead to an empty MRO for `Baz`, which is clearly wrong.

To break this cycle, it's important that the specializing `Foo[Baz]` should not be depending on `is_subset_eq()`. But this is impossible as long as we are still validating targs against tparam bound during specialization.

What this diff does is to make it such that while we are computing the class base types (which is needed when computing the class MRO), don't perform the bound validation. This does mean that we'll no longer be able to catch bound errors in base class list for now -- a problem that I'll fix in followup (spoiler: we'll need separate bindings for those).

Implementation wise we could of course achieve the goal by threading through a boolean flag to `expr_untype()` or `expr_infer()` indicating whether we are processing the class base list or not. But I don't quite like that approach because adding more flags to the already-complicated `expr_untype()` or `expr_infer()` doesn't feel composable -- I'm concerned that having one function with many flag tweaks serving many downstream clients could end up with big bloated functions that everyone depends on, which tend to be really hard to decouple / refactor. Besides, I believe a lot of work done in `expr_untype()` or `expr_infer()` is irrelevant to class base list processing in the first place, so it feels cleaner to just write a special version of `expr_untype()`/`expr_infer()` just for base class list processing, with restricted capability so we don't have to worry about unnecessary coupling and affecting pre-existing logic.

Reviewed By: stroxler

Differential Revision: D80387573

fbshipit-source-id: 65a9355b97aaa426505dcac23bf3fee0b3543497