Enforce base class syntactical form in binding phase
Summary:
In class metadata calculation, we only recognize 3 different syntactical forms of base classes: `X`, `X.Y`, and `X[Y]` (this list is subject to extension in  the future). Any other syntactical form is considered to be invalid base classes.

However, class base type calculation is not subject to the same constraint -- it'll just happily `expr_untype()` whatever expressions that appears on the base class list, creating inconsistencies of understanding base classes between class MRO and class metadata.

This diff attempts to align both calculations, by enforcing the base class syntactical forms in the binding phase. This way, class metadata and class base type calculation would be able to share the same logic of deciding which exprs on the base class list is considered valid, and which are not.

This approach does restrict our capabilities in the sense that we could no longer parse arbitrary expressions on the base class list (e.g. things like `class Foo(some_func(x)): ...` would have no way to work in the future). I consider this a good thing though -- it means that we'll be able to clearly reject a large amount of dubious / dynamic code patterns, and we'll never run into cycles when calculating base classes.

Reviewed By: stroxler

Differential Revision: D80387571

fbshipit-source-id: 7ac4f964842bf6d3d5d01f0eb23b7678fc61c573