Make `Variable::Parameter` to more clearly capture out-of-band data passing
Summary:
I noticed that both in discussion of turning off container inference and when dealing with
one of the big nondeterminism fixes last week, it became relevant that we have Vars that
are being used in a divergent way.

Most Vars result from solver-time type analysis (quantified vars used for directly solving,
"contained" vars that represent unpinned types we'll try to pin later, recursive vars used
in cycle breaking to defer picking a concrete type).

But lambda parameters and function parameters are different: they are used for bindings
code to insta-create a type, while deferring choosing the type until a *different* binding
evaluates (either the function definition binding for functions or the lambda expression
itself, whose hint if any determines contextual typing, for lambdas).

We've been using `contained` for these, which muddles the intent a bit. Making it a concrete
variant of `Variable` seems better.

There's no functional change here, just distinguishing cases that use the same mechanical
logic but have different intent, and taking advantage of that to comment how it
works because this is one of those highly nonlocal interactions where I think
good comments are especially important to maintainability.

It may be the case that we should eventually make some actual behavior changes - for example,
should `is_subset_eq_var` pin parameter types when the `self.union` flag is set? But for
now I think we should just separate the cases clearly.

Reviewed By: samwgoldman

Differential Revision: D80475050

fbshipit-source-id: 83a7894e4f994c0a2419e1b03a29831d9156af33