find method definitions separately from functions
Summary:
the bug identified in the previous diff is caused by us interpreting the method as a function.

When clicking the identifier itself, we see it as a function:
```
[fbcode/pyrefly/pyrefly/lib/state/lsp.rs:1458:17] &metadata = Variable(
    Some(
        Function,
    ),
)
```
but when clicking the reference, we see it as an attribute:
```
[fbcode/pyrefly/pyrefly/lib/state/lsp.rs:1458:17] &metadata = Attribute(
    Name("f"),
)
```

in order to correctly find-references on all definitions, we should always treat it as an attribute.

fixes https://github.com/facebook/pyrefly/issues/914

Reviewed By: stroxler

Differential Revision: D80480464

fbshipit-source-id: 9ddc5bb60016ac368d0ba4aba21a26d841ec8039