Remove `special_base` from bindings
Summary: `special_base` is just a separate `BaseClass` that'll be inserted into the base class list for every class. Previously, the insertion happens in the very beginning of answer phase. But there's actually no strong reason why that has to be the case -- we could just as well perform the insertion in the binding phase, so that the answer phase no longer needs to worry about it. This way, we can eliminate `special_base` field in both class metadata binding and class base binding, saving some spaces in the bindings table.

Reviewed By: rchen152

Differential Revision: D80487860

fbshipit-source-id: da1827be3f646217b785e2aaefb6382638f22463